Add perspective and orthographic projection cameras
